Libretro maintenance
