+++ libretro/libretro-graph.c
@@ -3,9 +3,8 @@
 #include <stdarg.h>
 #include <math.h>
 
-#include "libretro.h"
 #include "libretro-core.h"
-#include "graph.h"
+#include "libretro-graph.h"
 
 #include "font.i"
