+++ libretro/libretro-vkbd.h
@@ -1,5 +1,5 @@
-#ifndef VKBD_H
-#define VKBD_H
+#ifndef LIBRETRO_VKBD_H
+#define LIBRETRO_VKBD_H
 
 extern void print_virtual_kbd(unsigned short int *pixels);
 extern int check_vkey(int x, int y);
@@ -224,4 +224,4 @@ static retro_vkeys vkeys[VKBDX * VKBDY * 2] = {
 
 };
 
-#endif
+#endif /* LIBRETRO_VKBD_H */
